[algorithm][자료구조] 패키지 구조 변경
