+++ study/array_and_list.md
