+++ study/prefix_sum.md
